extract edit API into dedicated module
