+++ src/ServiceControl/MessageFailures/Api/RetryMessages.cs
@@ -1,10 +1,7 @@
 ﻿namespace ServiceControl.MessageFailures.Api
 {
-    using System;
     using System.Collections.Generic;
-    using System.IO;
     using System.Linq;
-    using System.Text;
     using InternalMessages;
     using Nancy;
     using Nancy.ModelBinding;
@@ -26,40 +23,6 @@
                 }
 
                 return await RetryMessagesApi.Execute(this, failedMessageId);
-            };
-
-            Post["/errors/{messageid}/editandretry", true] = async (parameters, token) =>
-            {
-                string failedMessageId = parameters.MessageId;
-
-                if (string.IsNullOrEmpty(failedMessageId))
-                {
-                    return HttpStatusCode.BadRequest;
-                }
-
-                //TODO: consider sending base64 encoded body from the client
-                string body;
-                using (var streamReader = new StreamReader(this.Request.Body))
-                {
-                    body = await streamReader.ReadToEndAsync().ConfigureAwait(false);
-
-                    if (string.IsNullOrWhiteSpace(body))
-                    {
-                        return HttpStatusCode.BadRequest;
-                    }
-                }
-
-                var base64String = Convert.ToBase64String(Encoding.UTF8.GetBytes(body));
-                await Bus.SendLocal(new EditAndSend
-                {
-                    FailedMessageId = failedMessageId,
-                    // Encode the body in base64 so that the new body doesn't have to be escaped
-                    NewBody = base64String,
-                }).ConfigureAwait(false);
-
-                
-
-                return HttpStatusCode.Accepted;
             };
 
             Post["/errors/retry", true] = async (_, token) =>
